requirements.txt
anthropic>=0.40.0
streamlit>=1.38.0
plotly>=5.18.0
pandas>=2.0.0
python-docx>=1.1.0
streamlit-authenticator>=0.4.0
pyyaml>=6.0
requests>=2.31.0
bcrypt>=4.0.0
